+++ TempEval/utils/__init__.py
@@ -1,4 +1,5 @@
 from ._github_export import export_to_github
 from ._file_handler import FileHandler
+from ._output_validation import sanitise_response
 
-__all__ = ['export_to_github', 'FileHandler']
+__all__ = ['export_to_github', 'FileHandler','sanitise_response']
