+++ pyproject.toml
@@ -29,3 +29,6 @@ dependencies = [
     'tqdm >= 4.67.1',
     'rogue-score == 0.1.2'
 ]
+[tool.setuptools.packages.find]
+include = ["tempeval*"]
+exclude = ["tempeval.tests*"]  # exclude tests from build artifacts
